A tool to rewrite git history, putting binary files under gitshed.
This allows existing repos to get the full benefit of gitshed.

The tool detects binary files based on file suffixes and content,
copies their content into the shed, and replaces them with symlinks
to that content.

TODO: Tests.
TODO: Tooling to upload the historical binary files to the content store.